chatbot/requirements.txt
uvicorn
fastapi
python-multipart 
jinja2
langchain==0.3.26
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone==0.2.8 
langchain-openai==0.3.24
langchain_ollama
langchain-huggingface>=0.1.0
langchain-community==0.3.26
-e .